DevOps: Build, Test and Deploy actions
Build and Test actions are run when a pull request is issued on any branch, and a push is done on any branch except 'trunk'. 'trunk' is excluded because of protections, so we have to create a push request to merge into trunk, and therefore it calls the action anyway.
The Deploy action is called when a push to the 'trunk' branch is made, so when a push request has been merged. It deploys the project to AppEngine