API: Starting TDD process
I decided to use a Test-Driven Development process to incrementally add features to the API. For the moment test users are returned, but the first tests have passed.
The strategy is to first define tests for the to-be implemented route, then develop the route handler. This allows to still pass tests as none are defined for not yet defined routes